Move usage listener tests to core from software-base
