+++ core/src/test/java/org/apache/brooklyn/core/mgmt/internal/LocalUsageManagerTest.java
@@ -16,67 +16,62 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.brooklyn.entity.software.base.test.core.mgmt.usage;
+
+package org.apache.brooklyn.core.mgmt.internal;
 
 import static org.testng.Assert.assertTrue;
 
 import java.util.List;
 
+import org.apache.brooklyn.api.location.Location;
+import org.apache.brooklyn.core.internal.BrooklynProperties;
+import org.apache.brooklyn.core.mgmt.usage.RecordingUsageListener;
+import org.apache.brooklyn.core.mgmt.usage.UsageListener;
+import org.apache.brooklyn.core.mgmt.usage.UsageManager;
+import org.apache.brooklyn.core.test.BrooklynAppUnitTestSupport;
+import org.apache.brooklyn.core.test.entity.LocalManagementContextForTests;
+import org.apache.brooklyn.core.test.entity.TestApplication;
 import org.apache.brooklyn.test.Asserts;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
-import org.apache.brooklyn.api.location.Location;
-import org.apache.brooklyn.core.entity.Entities;
-import org.apache.brooklyn.core.internal.BrooklynProperties;
-import org.apache.brooklyn.core.mgmt.internal.ManagementContextInternal;
-import org.apache.brooklyn.core.mgmt.usage.UsageListener;
-import org.apache.brooklyn.core.mgmt.usage.UsageManager;
-import org.apache.brooklyn.core.test.entity.LocalManagementContextForTests;
-import org.apache.brooklyn.core.test.entity.TestApplication;
 
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.Iterables;
 import com.google.common.collect.Lists;
 
-public class UsageListenerTest {
+public class LocalUsageManagerTest extends BrooklynAppUnitTestSupport {
 
     // Also see {Application|Location}UsageTrackingTest for listener functionality
-    
-    private static final Logger LOG = LoggerFactory.getLogger(ApplicationUsageTrackingTest.class);
 
-    protected TestApplication app;
-    protected ManagementContextInternal mgmt;
-
+    @Override
     protected boolean shouldSkipOnBoxBaseDirResolution() {
         return true;
     }
 
+    @Override
     @BeforeMethod(alwaysRun=true)
     public void setUp() throws Exception {
         RecordingStaticUsageListener.clearInstances();
+        super.setUp();
     }
 
+    @Override
     @AfterMethod(alwaysRun=true)
     public void tearDown() throws Exception {
-        try {
-            if (mgmt != null) Entities.destroyAll(mgmt);
-        } catch (Throwable t) {
-            LOG.error("Caught exception in tearDown method", t);
-        } finally {
-            mgmt = null;
-            RecordingStaticUsageListener.clearInstances();
-        }
+        super.tearDown();
+        RecordingStaticUsageListener.clearInstances();
     }
 
     @Test
-    public void testAddUsageListenerViaProperties() throws Exception {
+    public void testAddUsageListenerInstance() throws Exception {
         BrooklynProperties brooklynProperties = BrooklynProperties.Factory.newEmpty();
         brooklynProperties.put(UsageManager.USAGE_LISTENERS, RecordingStaticUsageListener.class.getName());
         mgmt = LocalManagementContextForTests.newInstance(brooklynProperties);
-        
+        assertUsageListenerCalledWhenApplicationStarted();
+    }
+
+    private void assertUsageListenerCalledWhenApplicationStarted() {
         app = TestApplication.Factory.newManagedInstanceForTests(mgmt);
         app.start(ImmutableList.<Location>of());
 
@@ -86,10 +81,10 @@ public class UsageListenerTest {
                 assertTrue(events.size() > 0, "events="+events); // expect some events
             }});
     }
-    
+
     public static class RecordingStaticUsageListener extends RecordingUsageListener implements UsageListener {
         private static final List<RecordingStaticUsageListener> STATIC_INSTANCES = Lists.newCopyOnWriteArrayList();
-        
+
         public static RecordingStaticUsageListener getInstance() {
             return Iterables.getOnlyElement(STATIC_INSTANCES);
         }
@@ -97,7 +92,7 @@ public class UsageListenerTest {
         public static void clearInstances() {
             STATIC_INSTANCES.clear();
         }
-        
+
         public RecordingStaticUsageListener() {
             // Bad to leak a ref to this before constructor finished, but we'll live with it because
             // it's just test code!
